+++ src/templater/Template.java
@@ -1,9 +1,7 @@
 package templater;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.Arrays;
 
-import templater.compiler.LanguageElement;
 import templater.compiler.parser.Parser;
 import templater.compiler.tokeniser.Tokeniser;
 import templater.compiler.tokeniser.UnknownTokenException;
@@ -28,11 +26,9 @@ class Template {
    *               details this template.
    */
   public Template(String source) throws UnknownTokenException {
-    List<LanguageElement> l = new ArrayList<>();
-    l.add(
+    this.syntaxTree = new Root(Arrays.asList(
       new Parser(new Tokeniser(source).tokenise()).parse()
-    );
-    this.syntaxTree = new Root(l);
+    ));
   }
 
   /**
