+++ src/templater/compiler/parser/ClassAttributeMatcher.java
@@ -1,7 +1,7 @@
 package templater.compiler.parser;
 
-import templater.compiler.tokeniser.Token;
 import templater.language.StringResolvables;
+import templater.language.Token;
 
 class ClassAttributeMatcher extends TokenMatchable<StringResolvables> {
   @Override
