+++ src/templater/compiler/tokeniser/RegexMatcher.java
@@ -3,7 +3,8 @@ package templater.compiler.tokeniser;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import templater.compiler.TextFilePosition;
+import templater.language.Token;
+import templater.language.TokenKind;
 
 abstract class RegexMatcher extends TokenMatcher {
   private final Pattern regex;
