+++ src/templater/compiler/tokeniser/StringLiteralMatcher.java
@@ -1,5 +1,7 @@
 package templater.compiler.tokeniser;
 
+import templater.language.TokenKind;
+
 class StringLiteral extends EscapedDelimitedMatcher {
   StringLiteral() {
     super("\"", '"', TokenKind.STRING_LITERAL);
