+++ src/templater/compiler/tokeniser/TokenMatcher.java
@@ -1,6 +1,7 @@
 package templater.compiler.tokeniser;
 
 import templater.compiler.Matchable;
+import templater.language.Token;
 
 abstract class TokenMatcher implements Matchable<CharListQueue, Token> {
 }
