+++ src/templater/language/LanguageElement.java
@@ -1,4 +1,4 @@
-package templater.compiler;
+package templater.language;
 
 public class LanguageElement {
 }
