+++ src/templater/language/Node.java
@@ -4,8 +4,6 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import templater.compiler.LanguageElement;
-
 /**
  * A class representing a node in the syntax tree.
  *
